[SWM-269] Refactor : constants directory
