--- src/constants/lectureRecommendations/lectureRecommendations.ts
@@ -1,5 +1,0 @@
-import { ONE_MINUTE_IN_MS } from "../auth/auth";
-
-// recommendation query
-export const CACHE_TIME = ONE_MINUTE_IN_MS * 30;
-export const STALE_TIME = ONE_MINUTE_IN_MS;
