+++ src/constants/query/query.ts
@@ -1,0 +1,4 @@
+import { ONE_MINUTE_IN_MS } from "../time/time";
+
+export const CACHE_TIME = ONE_MINUTE_IN_MS * 30;
+export const STALE_TIME = ONE_MINUTE_IN_MS * 5;
