--- src/constants/scheduling/scheduling.ts
@@ -1,2 +1,0 @@
-export const FOUR_HOURS_IN_MINUTES = 240;
-export const THIRTY_MINUTES = 30;
