--- src/constants/search/search.ts
@@ -1,6 +1,0 @@
-import { ONE_MINUTE_IN_MS } from "../auth/auth";
-
-//search query
-export const LIMIT_PER_FETCH = 20;
-export const CACHE_TIME = ONE_MINUTE_IN_MS * 30;
-export const STALE_TIME = ONE_MINUTE_IN_MS * 5;
