+++ src/constants/skeleton/skeleton.ts
@@ -1,0 +1,6 @@
+// index, review query
+export const INDEX_LIMIT = 5;
+export const REVIEW_LIMIT = 10;
+
+// classroom query
+export const CLASSROOM_LIMIT = 8;
