+++ src/constants/time/time.ts
@@ -3,5 +3,9 @@ export const ONE_HOUR = 60 * 60;
 export const ONE_WEEK = 7 * 24 * ONE_HOUR;
 export const SESSION_AGE = ONE_WEEK - ONE_HOUR;
 
-//refresh
+//silent refresh
 export const ONE_MINUTE_IN_MS = 60 * 1000;
+
+//scheduling
+export const FOUR_HOURS_IN_MINUTES = 240;
+export const THIRTY_MINUTES = 30;
