--- src/constants/ui/toast.ts
@@ -1,2 +1,0 @@
-export const TOAST_TIMEOUT = 4 * 1000;
-export const TOAST_DELAY = 4;
