+++ src/util/toast/index.ts
@@ -1,0 +1,1 @@
+export const TOAST_TIMEOUT = 4 * 1000;
